invoices-mcp/pyproject.toml
[project]
name = "invoices-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.5.5",
    "fastmcp>=2.10.6",
    "mistralai>=1.9.2",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
]
